+++ src/main/java/me/srrapero720/watermedia/api/player/FirePlayer.java
@@ -36,7 +36,7 @@ public class FirePlayer extends AbstractPlayer {
     }
 
     @Override
-    public void seekFastTo(int ticks) {
+    public void seekFastTo(long ticks) {
 
     }
